+++ Common/Snapshot/ObstacleDto.cpp
@@ -1,9 +1,9 @@
 #include "ObstacleDto.h"
 
-ObstacleDto::ObstacleDto(const uint8_t& id, const TypeObstacle& type,
+ObstacleDto::ObstacleDto(const uint16_t& id, const TypeObstacle& type,
         const std::pair<int16_t, int16_t>& position) : id(id), type(type), position(position) {}
 
-uint8_t ObstacleDto::getId() const {
+uint16_t ObstacleDto::getId() const {
     return id;
 }
