+++ Common/Snapshot/ObstacleDto.h
@@ -7,16 +7,16 @@
 #include "TypeObstacle.h"
 
 class ObstacleDto {
-    uint8_t id;
+    uint16_t id;
     TypeObstacle type;
     std::pair<int16_t, int16_t> position;
     
     public:
 
-    explicit ObstacleDto(const uint8_t& id, const TypeObstacle& type,
+    explicit ObstacleDto(const uint16_t& id, const TypeObstacle& type,
         const std::pair<int16_t, int16_t>& position);
 
-    uint8_t getId() const;
+    uint16_t getId() const;
 
     TypeObstacle getTypeObstacle() const;
